Unified Stage code and reviewed game code. Will merge if no issues are found.
