--- client/src/pages/Stage1.css
@@ -1,9 +1,0 @@
-/* styles */
-.stage1_game_section {
-  background-image: url("../assets/img_background/background1.webp");
-  background-size: cover;
-  min-height: 1200px;
-  height: 100vh;
-  width: 100vw;
-  padding-top: 5px;
-}
