--- client/src/pages/Stage2.css
@@ -1,9 +1,0 @@
-/* styles */
-.stage2_game_section {
-  background-image: url("../assets/img_background/background3.webp");
-  background-size: cover;
-  min-height: 1200px;
-  height: 100vh;
-  width: 100vw;
-  padding-top: 5px;
-}
